refactor(Container): move Stream to MainColumn
